+++ modules-js/graphql-typescript/package.json
@@ -20,7 +20,7 @@
     "@cityofboston/config-jest-typescript": "^0.0.0",
     "@cityofboston/config-typescript": "^0.0.0",
     "@types/jest": "23.x.x",
-    "@types/node": "^10.0.0",
+    "@types/node": "^8.0.0",
     "jest": "23.6.0",
     "rimraf": "^2.6.2",
     "ts-jest": "23.10.0",
