+++ modules-js/hapi-next/package.json
@@ -27,7 +27,7 @@
     "@cityofboston/config-typescript": "^0.0.0",
     "@types/hapi": "^17.0.19",
     "@types/jest": "23.x.x",
-    "@types/node": "^10.0.0",
+    "@types/node": "^8.0.0",
     "hapi": "^17.4.0",
     "jest": "23.6.0",
     "rimraf": "^2.6.2",
