+++ modules-js/mssql-common/package.json
@@ -24,7 +24,7 @@
     "@cityofboston/config-typescript": "^0.0.0",
     "@types/jest": "23.x.x",
     "@types/mssql": "^4.0.10",
-    "@types/node": "^10.0.0",
+    "@types/node": "^8.0.0",
     "jest": "23.6.0",
     "rimraf": "^2.6.2",
     "ts-jest": "23.10.0",
