+++ modules-js/mssql-typescript/package.json
@@ -19,7 +19,7 @@
   "devDependencies": {
     "@cityofboston/config-typescript": "^0.0.0",
     "@types/knex": "^0.14.24",
-    "@types/node": "^10.0.0",
+    "@types/node": "^8.0.0",
     "rimraf": "^2.6.2",
     "tsc-watch": "^1.0.26",
     "typescript": "^3.0.0"
