+++ services-js/access-boston/package.json
@@ -81,7 +81,7 @@
     "@storybook/addons": "4.0.0-alpha.22",
     "@storybook/react": "4.0.0-alpha.22",
     "@types/jest": "^22.2.3",
-    "@types/node": "^9.6.5",
+    "@types/node": "^8.0.0",
     "@types/react": "^16.4.7",
     "@types/storybook__addon-actions": "^3.0.3",
     "@types/storybook__addon-storyshots": "^3.4.1",
