+++ services-js/commissions-app/package.json
@@ -73,7 +73,7 @@
     "@types/isomorphic-fetch": "^0.0.34",
     "@types/jest": "^22.2.3",
     "@types/mssql": "^4.0.10",
-    "@types/node": "^9.6.5",
+    "@types/node": "^8.0.0",
     "@types/react": "^16.4.7",
     "@types/storybook__addon-actions": "^3.0.3",
     "@types/storybook__addon-storyshots": "^3.4.1",
