+++ services-js/registry-certs/.env.sample
@@ -1,10 +1,10 @@
 API_KEYS=
 WEB_API_KEY=
-REGISTRY_DB_USER=
-REGISTRY_DB_PASSWORD=
-REGISTRY_DB_DOMAIN=
-REGISTRY_DB_SERVER=
-REGISTRY_DB_DATABASE=
+REGISTRY_DATA_DB_USER=
+REGISTRY_DATA_DB_PASSWORD=
+REGISTRY_DATA_DB_DOMAIN=
+REGISTRY_DATA_DB_SERVER=
+REGISTRY_DATA_DB_DATABASE=
 
 STRIPE_PUBLISHABLE_KEY=
 STRIPE_SECRET_KEY=
