+++ services-js/registry-certs/scripts/generate-fixtures.ts
@@ -6,30 +6,31 @@ import fs from 'fs';
 import path from 'path';
 import faker from 'faker';
 
-import { makeRegistryDataFactory } from '../server/services/RegistryData';
+import { makeRegistryDbFactory } from '../server/services/RegistryDb';
+import { DatabaseConnectionOptions } from '@cityofboston/mssql-common';
 
 dotenv.config();
 
 const FIXTURE_SEARCHES = ['smith'];
 
 (async function generateFixtures() {
-  const registryFactoryOpts = {
-    user: process.env.REGISTRY_DATA_DB_USER,
-    password: process.env.REGISTRY_DATA_DB_PASSWORD,
+  const registryFactoryOpts: DatabaseConnectionOptions = {
+    username: process.env.REGISTRY_DATA_DB_USER!,
+    password: process.env.REGISTRY_DATA_DB_PASSWORD!,
     domain: undefined,
     server: process.env.REGISTRY_DATA_DB_SERVER,
     database: process.env.REGISTRY_DATA_DB_DATABASE,
   };
 
-  const registryDataFactory = await makeRegistryDataFactory(
+  const registryDbFactory = await makeRegistryDbFactory(
     null as any,
     registryFactoryOpts
   );
-  const registryData = registryDataFactory.registryData();
+  const registryDb = registryDbFactory.registryDb();
 
   await Promise.all(
     FIXTURE_SEARCHES.map(async search => {
-      const records = await registryData.search(search, 0, 500, null, null);
+      const records = await registryDb.search(search, 0, 500, null, null);
 
       const cleanedRecords = records.map(r => {
         const firstName = faker.name.firstName().toUpperCase();
@@ -50,7 +51,7 @@ const FIXTURE_SEARCHES = ['smith'];
     })
   );
 
-  await registryDataFactory.cleanup();
+  await registryDbFactory.cleanup();
 })().catch(err => {
   console.error(err);
   process.exit(1);
