+++ services-js/registry-certs/server/stripe-events.test.ts
@@ -1,7 +1,6 @@
 import { processStripeEvent } from './stripe-events';
 
-import RegistryData from './services/RegistryData';
-import RegistryOrders from './services/RegistryOrders';
+import RegistryDb from './services/RegistryDb';
 
 import CHARGE_SUCCEEDED from '../fixtures/stripe/charge-succeeded.json';
 import ORDER from '../fixtures/registry-orders/order.json';
@@ -10,18 +9,15 @@ import CERTIFICATES from '../fixtures/registry-data/smith.json';
 describe('charge.created', () => {
   let emails;
   let stripe;
-  let registryData: RegistryData;
-  let registryOrders: RegistryOrders;
+  let registryDb: RegistryDb;
 
   beforeEach(() => {
     emails = {
       sendReceiptEmail: jest.fn(),
     } as any;
     stripe = {} as any;
-    registryData = {
+    registryDb = {
       lookup: () => CERTIFICATES[0],
-    } as any;
-    registryOrders = {
       findOrder: () => ORDER,
       addPayment: jest.fn(),
     } as any;
@@ -32,15 +28,14 @@ describe('charge.created', () => {
       {
         emails,
         stripe,
-        registryData,
-        registryOrders,
+        registryDb,
       },
-      null,
+      '',
       '',
       JSON.stringify(CHARGE_SUCCEEDED)
     );
 
-    expect(registryOrders.addPayment).toHaveBeenCalledWith(
+    expect(registryDb.addPayment).toHaveBeenCalledWith(
       19, // the order key from the charge as an int
       expect.anything(),
       expect.anything(),
@@ -53,10 +48,9 @@ describe('charge.created', () => {
       {
         emails,
         stripe,
-        registryData,
-        registryOrders,
+        registryDb,
       },
-      null,
+      '',
       '',
       JSON.stringify(CHARGE_SUCCEEDED)
     );
