+++ templates/js-server-module/template/package.json
@@ -20,7 +20,7 @@
   "devDependencies": {
     "@cityofboston/config-jest-typescript": "^0.0.0",
     "@cityofboston/config-typescript": "^0.0.0",
-    "@types/node": "^10.0.0",
+    "@types/node": "^8.0.0",
     "@types/jest": "23.x.x",
     "jest": "23.6.0",
     "rimraf": "^2.6.2",
